filth_projections: fix rays going through walls
This would happen rarely if the is_inside_tile check failed, but the ray
started close enough to the tile that it didn't register a collision.

To fix this, I cast the ray from the dust towards the node (instead of
the other way around), because we can assume that the starting point of
the dust is not inside a tile. (This is technically not true, but I'm
not super interested in cases where dust is on the internal edge of a
block of tiles.)